(refactor) Change the grammar for custom format strings and simplify the parser.
In preparation for further refactors, the parser of a format string
has been moved to a new FormatParser struct and much of the parsing
logic has been simplified and consolidated.

Additionally, the grammar of the format string has been changed
to close attribute applications with a single `@` rather than `@@`
as this allows the parser to be unambiguous with only single
lookahead.