(feat) Add a prelude module that is meant to be glob imported.
This module includes many traits which only provide methods
on existing types, as well as the most-used concrete types in
Iron.

The hope is that use of this module will lessen confusion about
"no such method .set" and related errors from new Iron users
who are not yet aware of the use of extension traits.