(feat) Stream the body of a Request as a Reader
This changes the behavior of Iron in a rather crucial way.
Instead of reading the entire body of an incoming Request into a
`Vec<u8>` at the creation time of a Request, the body is now
exposed as a Reader on the Request object and can be read on-demand
or not-at-all depending on its use.

Streaming is critical for uses such as file uploading, where loading
the entire request body into memory unconditionally is... bad.

The body is exposed as a mostly opaque `Body` type which implements
`Reader` and wraps hyper's `HttpReader<&'a mut (Reader + 'a)>`. The http
reader cannot currently be retrieved from a `Body`, but for testing
purposes `Body::new` can be used to construct a `Body` from an `HttpReader`.

Progress on #263
Fixes #265