(feat) Rewrite IronError to wrap an Error and a Response.
The new IronError wraps the old IronError (Box<Error>) and a Response. The
old IronError filled only one of two important roles; it old informed
receivers what went wrong, but did not tell them what to about it. The new
IronError uses an embedded Response to also inform receivers what the
appropriate reaction to the error is.

At the end of the new error-handling flow the Response embedded in the
IronError will be used as the Response that is sent to the client. This allows
the thrower of the error to make a suggestion as to what should happen in the
event of an error, but gives downstream receivers the chance to override it.

[breaking-change]

Part of error simplification (#272)
Progress towards ChainBuilder fixes (#186)