(feat) Rewrite the error handling parts of the middleware system.
This commit completely revamps the error-handling and receiving parts of Iron
to use a totally new system based on the new IronError type. The error flow is
*vastly* simplified with the removal of "handling", replaced with the ability
to set the Response to be sent as part of the error.

The flow is now very simple to describe - if at any point an error is returned,
the middleware flow switches to failure mode, where it proceeds as usual
except it calls the `catch` methods of Before and After middleware instead of
their `before` and `after` methods. At the end of the flow, the Response
remaining on the IronError instance is sent to the client. That's it!

Fixes #272
Fixes #186