(feat) Add ability to serve over HTTPS.
This adds a protocol argument to Iron::listen_with, allowing the user
to specify HTTPS as the protocol instead of HTTP. This property can
later be extended with more protocols, such as HTTP2. Iron::listen
is replaced with two methods, http(addr) and https(addr, certificate, key).
These methods act as convenience methods to create a listener of the
appropriate type.