Switch response abstraction from `Read` to `WriteBody`.
A `Read`er is not the best abstraction for the response body - it has to
be copied to the output `Write`r through an extra intermediary buffer
and complicates the implementation for users that want to compute the
response on the fly.

This design is modeled after Java JAX-RS's StreamingOutput class. The
`write_body` method takes the Hyper `Response` object concretely instead
of a generic `Write`r to allow for future optimizations with Hyper, for
example, allowing the underlying file descriptor to be retrieved for use
with `sendfile` on Linux.

This is a [breaking-change]