(refactor) Add a new public listening function for custom Server
This allows consumers to set up the server as desired, such as configuring the
SSL context for client certificates.

The API is a little iffy, since consumers still need to pass in an instance of
Protocol with certificate and key, which will be unused.

Fixes #447, and forms a starting point for further abstraction in #261.