Bring rust-url's PathFormatter implementation in-tree
In version 1.0.0 of rust-url, the PathFormatter has been removed. I
brought its implementation[1] over since it wasn't much code.

The check for the vector being empty might be unnecessary since the
documentation at the top of this file says that path is "A *non-empty*
vector encoding the parts of the URL path", but I wasn't sure how
strongly that guarantee was upheld.

[1] -
https://github.com/servo/rust-url/commit/9e759f18726c8e1343162922b87163d4dd08fe3c#diff-0bb16ac13b75e9b568fa4aff61b0e71dL24