README: Add note about maintenance status of iron (#576)
The last release of iron (0.6.0) took months to do and required a lot of pushing
from an outsider (me), see #551. Given that, I think it's fair to say that iron
is unmaintained and at this point I would advise new people not to use it.

If any of that changes, we can remove the notice again.